requirements.txt
discord==1.7.3
discord.py==1.7.3
discord.py[voice]
discord_webhook==1.3.0
discord-ext-bot==1.0.1
python-dotenv==1.0.0
discord-py-slash-command==2.0.0
mysql-connector-python==8.2.0
openai==0.28